Move necessary data structures into the strato-model group
